feat(gallery): add users only if they do not exist
add users if they do not exist

MODULE-2-PART-4-mongodb
