+++ backend/src/models/image.model.ts
@@ -1,0 +1,8 @@
+import mongoose, { Schema } from 'mongoose';
+
+const ImagesSchema: Schema = new Schema({
+  path: {type: String, required: true, unique: true},
+  metadata: {type: Object, required: true}
+});
+
+export const images = mongoose.model('Image', ImagesSchema);
