+++ backend/src/services/db-service.ts
@@ -1,13 +1,6 @@
-import mongoose, { Schema } from 'mongoose';
 import { stat } from 'node:fs/promises';
-import { new_users } from '../users/users.js';
-
-const ImagesSchema: Schema = new Schema({
-  path: {type: String, required: true, unique: true},
-  metadata: {type: Object, required: true}
-});
-
-
+import { users } from '../models/user.model.js';
+import { images } from '../models/image.model.js';
 
 export async function uploadImageDataToDB(req) {
 
@@ -27,8 +20,36 @@ export async function uploadImageDataToDB(req) {
 }
 
 export async function findUserInDB(email) {
-  const user = await new_users.findOne({email: email}, {_id: 0, __v: 0} ).exec();
+  const user = await users.findOne({email: email}, {_id: 0, __v: 0} ).exec();
   return user;
 }
 
-export const images = mongoose.model('Image', ImagesSchema);
+export async function addDefaultUsersToDB() {
+  const defaultUsersArray = [
+    'asergeev@flo.team',
+    'tpupkin@flo.team',
+    'vkotikov@flo.team',
+  ];
+
+  const records = await users.find({ 'email': { $in: defaultUsersArray } });
+
+  if (records.length !== 0) return;
+
+  const asergeev = new users({
+    email: 'asergeev@flo.team',
+    password: 'jgF5tn4F',
+  });
+  asergeev.save().then(() => console.log('asergeev was saved'));
+
+  const tpupkin = new users({
+    email: 'tpupkin@flo.team',
+    password: 'tpupkin@flo.team',
+  });
+  tpupkin.save().then(() => console.log('tpupkin was saved'));
+
+  const vkotikov = new users({
+    email: 'vkotikov@flo.team',
+    password: 'po3FGas8',
+  });
+  vkotikov.save().then(() => console.log('vkotikov was saved'));
+}
